requirement.txt
ImageHash==4.3.1
numpy==1.24.2
Pillow==9.4.0
PyWavelets==1.4.1
scipy==1.10.0
vptree==1.3
